+++ src/frontend-common/game_list.h
@@ -129,7 +129,7 @@ private:
   enum : u32
   {
     GAME_LIST_CACHE_SIGNATURE = 0x45434C47,
-    GAME_LIST_CACHE_VERSION = 27
+    GAME_LIST_CACHE_VERSION = 28
   };
 
   using CacheMap = std::unordered_map<std::string, GameListEntry>;