+++ src/frontend-common/game_settings.h
@@ -61,6 +61,8 @@ struct Entry
   std::optional<bool> display_linear_upscaling;
   std::optional<bool> display_integer_upscaling;
   std::optional<bool> display_force_4_3_for_24bit;
+  std::optional<u16> display_aspect_ratio_custom_numerator;
+  std::optional<u16> display_aspect_ratio_custom_denominator;
   std::optional<u32> gpu_resolution_scale;
   std::optional<u32> gpu_multisamples;
   std::optional<bool> gpu_per_sample_shading;